1684 split holdermodule (#653)
* split holder module to smaller modules to improve maintainability

* cleanup
