+++ holder/src/main/java/nl/rijksoverheid/ctr/holder/modules/ErrorsModule.kt
@@ -1,0 +1,24 @@
+package nl.rijksoverheid.ctr.holder.modules
+
+import nl.rijksoverheid.ctr.api.factory.NetworkRequestResultFactory
+import nl.rijksoverheid.ctr.shared.factories.ErrorCodeStringFactory
+import nl.rijksoverheid.ctr.shared.factories.ErrorCodeStringFactoryImpl
+import org.koin.core.qualifier.named
+import org.koin.dsl.module
+
+/*
+ *  Copyright (c) 2021 De Staat der Nederlanden, Ministerie van Volksgezondheid, Welzijn en Sport.
+ *   Licensed under the EUROPEAN UNION PUBLIC LICENCE v. 1.2
+ *
+ *   SPDX-License-Identifier: EUPL-1.2
+ *
+ */
+val errorsModule = module {
+    factory {
+        NetworkRequestResultFactory(get(named("ResponseError")))
+    }
+
+    factory<ErrorCodeStringFactory> {
+        ErrorCodeStringFactoryImpl()
+    }
+}
